final version with server crtl+c control

+++ src/global.cpp
@@ -4,19 +4,18 @@
 #include <sstream>
 #include <iomanip>
 
-std::string create_header(int command_type, size_t message_length) {
+std::string create_header(int command_type, int message_length) {
     std::ostringstream oss;
     oss << std::setw(10) << std::setfill('0') << message_length;
     std::string padded_length = oss.str();
 
     std::string header;
     std::string div = ":";
-    // CMD:00LEN|msg1:msg2:msg3
     header = std::to_string(command_type) + div + padded_length + "|";
     return header;
 }
 
-bool parseHeader(const std::string& header, MsgType& command, size_t& message_length) {    
+bool parseHeader(const std::string& header, MsgType& command, int& message_length) {    
     int div_pos =  header.find(':');
     int type = std::stoi(header.substr(0, div_pos));
     command = (MsgType) type;
@@ -39,7 +38,7 @@ void string_procent_encode(std::string& s){
 
 void string_procent_decode(std::string& s){
     std::string decoded; 
-    for (int i = 0; i < s.length(); ++i) {
+    for (size_t i = 0; i < s.length(); ++i) {
         if (s[i] == '%' && s[i+1] == '3' && s[i+2] == 'A') {
             decoded += ':';
             i += 2;
